refactor DnsFuture out of dns_handle
